implemented line-number lookup
Added API to output which file/line the current instruction is for.

Implemented the line program machine according to the dwarf 4 spec until
we can process our debugee program successfully. Libdwarf does this for
us, but it is more fun and interesting to do it ourselves, so we do

Implementation details:
+ Current implementation is pretty naive, it reruns the whole line
  program for a compilation unit every time a source location is looked
  up. This is easily optimisable in the future (save machine state for
  each fn or something), but for now is good enough.
+ Track compilation units to figure out which line program to run